Modulo15 - separando lógica do resto dos controller em services
