+++ packages/GoBarber/backend/src/app/controllers/AppointmentController.js
@@ -1,11 +1,8 @@
-import { isBefore, subHours } from 'date-fns';
-
 import Appointment from '../models/Appointment';
 import User from '../models/User';
 import File from '../models/File';
-import Queue from '../../lib/Queue';
-import CancellationMail from '../jobs/CancellationMail';
 import CreateAppointmentService from '../services/CreateAppointmentService';
+import CancelAppointmentService from '../services/CancelAppointmentService';
 
 class AppointmentController {
   async index(req, res) {
@@ -48,39 +45,10 @@ class AppointmentController {
   }
 
   async delete(req, res) {
-    const appointment = await Appointment.findByPk(req.params.id, {
-      include: [
-        {
-          model: User,
-          as: 'provider',
-          attributes: ['name', 'email'],
-        },
-        {
-          model: User,
-          as: 'user',
-          attributes: ['name'],
-        },
-      ],
+    const appointment = await CancelAppointmentService.run({
+      providerId: req.params.id,
+      userId: req.userId,
     });
-
-    if (appointment.userId !== req.userId) {
-      return res.status(401).json({
-        error: "You don't have permission to cancel this appointment.",
-      });
-    }
-
-    const dateSubHours = subHours(appointment.date, 2);
-    if (isBefore(dateSubHours, new Date())) {
-      return res.status(401).json({
-        error: 'You can only cancel appointments 2 hours in advance.',
-      });
-    }
-
-    appointment.canceledAt = new Date();
-
-    await appointment.save();
-
-    await Queue.add(CancellationMail.key, { appointment });
 
     return res.json(appointment);
   }
